Add a gitlab-ci job for the unix Makefile.
Ensure test coverage for the non-autotools makefile. This is
much faster than the autotools build, but doesn't currently
support tests or a `dist` target.

The syntax of the travis-ci.org build description file makes
it tedious to add alternate build types. It's much easier with
gitlab, so it's nice to have testing outside of jenkins.