Fix documentation Makefile.
We ship a default Makefile which can be used to invoke doxygen
to format the API documentation outside of the autotools build,
which requires implementing subsitition of the appropriate
variables in Doxyfile.in.

This commit adds a missing substitution, telling doxygen to
expect `dot` from the graphviz package to be available.